lib/storage: use lrucache to implement tagFilters loops cache
The tagFilters loop cache is per-indexDB which means that currently there are
two instances, one for idbCurr and one for idbPrev. When the partition index is
released, there will be as many instances of this cache as there will be
partitions.

The cache is implemented using workingsetcache. Which occupies at least 30MB
even when unused. Given that only the latest indexDB is used most of the time,
a lot of memory can be wasted.

Therefore the cache implementation is changed to lrucache because it does not
consume memory when it is unused and also has timeout-based eviction.

This is a follow-up for 4cd727a511 (https://github.com/VictoriaMetrics/VictoriaMetrics/pull/10072).

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>