app/vmauth: properly increment metric xxx_user_request_backend_errors_total
Currently, backendErrors may be counted twice if a request to the
backend fails due to context.DeadlineExceeded.

https://github.com/VictoriaMetrics/VictoriaMetrics/blob/9bc7a17d80b8fa8e497532b51cf25bbe36ef9c3f/app/vmauth/main.go#L328

https://github.com/VictoriaMetrics/VictoriaMetrics/blob/9bc7a17d80b8fa8e497532b51cf25bbe36ef9c3f/app/vmauth/main.go#L294

And we increment this counter in a way that is somewhat inconsistent.
Given that the counter's name is `xx_request_backend_errors_total`, it
should only increase when a backend request returns an error. This value
can exceed the user request error count if multiple backend requests
fail for a single user request.
The `xxx_request_backend_errors_total` counter should be used in
conjunction with the `xxx_request_backend_requests_total` introduced in
https://github.com/VictoriaMetrics/VictoriaMetrics/pull/10171.