lib/storage: fix per-idb cache stats
This fixes the following corner case: if all instances of a cache have zero
size, the stats won't be set at all. This results in some weird graphs if the
cache is reset very often (such as tfssCache): the cache sizeMaxBytes alternates
between the actual value and zero.

Follow-up for f62893c151

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>